Update: Rename to hub database from metadata
