+++ src/main/java/com/cmyk/ego/speaktoyouspring/config/HubConfig.java
@@ -1,6 +1,6 @@
 package com.cmyk.ego.speaktoyouspring.config;
 
-import com.cmyk.ego.speaktoyouspring.config.properties.MetadataProperties;
+import com.cmyk.ego.speaktoyouspring.config.properties.HubProperties;
 import jakarta.persistence.EntityManagerFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -19,14 +19,14 @@ import javax.sql.DataSource;
 @RequiredArgsConstructor
 @EnableJpaRepositories(
         // metadata 패키지 하위의 @Repository들을 대상으로 설정
-        basePackages = MetadataConfig.BASE_PACKAGE,
+        basePackages = HubConfig.BASE_PACKAGE,
         entityManagerFactoryRef = "metadataEntityManagerFactory", // 엔티티를 조작할 Manager를 생성
         transactionManagerRef = "metadataTransactionManager") // (클래스 내부에 작성되어 있는 함수명)
-public class MetadataConfig {
+public class HubConfig {
     // 파일 내에서만 url을 사용 가능하도록 Default 접근 지정자 설정
-    static final String BASE_PACKAGE = "com.cmyk.ego.speaktoyouspring.api.metadata";
+    static final String BASE_PACKAGE = "com.cmyk.ego.speaktoyouspring.api.hub";
 
-    private final MetadataProperties metadataProperties;
+    private final HubProperties hubProperties;
 
     /// DataSource와 관련된 Bean이 다양한데, Bean 할당 순서를 수동으로 조정하기 위함이다.
     /// DataSource를 사용하는 다른 클래스에선 @Qualifier를 사용한다.
@@ -35,7 +35,7 @@ public class MetadataConfig {
     @Primary
     @Bean
     public DataSource metadataDataSource() {
-        return metadataProperties.getDatasource();
+        return hubProperties.getDatasource();
     }
 
     /// 엔티티를 조작할 Manager를 생성
@@ -48,8 +48,8 @@ public class MetadataConfig {
         return builder
                 .dataSource(metadataDataSource()) // 데이터 소스 추가
                 // 조작할 클래스(경로 하위에 있는 @Repository) 명시
-                .packages(MetadataConfig.BASE_PACKAGE)
-                .persistenceUnit("metaDB") // 여러 개의 EntityManagerFactory를 구분하기 위한 논리적인 식별자
+                .packages(HubConfig.BASE_PACKAGE)
+                .persistenceUnit("hubDB") // 여러 개의 EntityManagerFactory를 구분하기 위한 논리적인 식별자
                 .build();
     }
