+++ src/main/resources/application.yml
@@ -18,38 +18,22 @@ spring:
     validate-migration-naming: true # 파일명 형식이 Flyway 규칙을 따르는지 검증
     baseline-on-migrate: true # Flyway가 버전이 지정되지 않은 기존 데이터베이스에서도 마이그레이션을 실행할 수 있도록 설정
     locations: # 마이그레이션 파일을 찾을 디렉토리를 지정
-      - classpath:db/migration/metadata
+      - classpath:db/migration/hub
       - classpath:db/migration/tenant
 
-  # cors error 해결
-  web:
-    cors:
-      allowed-origins: "http://localhost:8080"
-      allowed-methods: "GET,POST,PUT,DELETE,OPTIONS"
-      allowed-headers: "Authorization,Content-Type,X-TenantID"
-      exposed-headers: "Authorization"
-      allow-credentials: true
-
 # metadata database 환경설정
-metadata:
+hub:
   datasource:
-    jdbc-url: jdbc:postgresql://localhost:5432/metadata
+    jdbc-url: jdbc:postgresql://localhost:5432/hub
     username: user1
     password: user1234
-#    hikari:
-#      connection-test-query: SELECT 1
 
 # tenant database 환경설정
-
 tenant:
   datasource:
     jdbc-url: jdbc:postgresql://localhost:5432/tenant
     username: user1
     password: user1234
-#    hikari:
-#      connection-test-query: SELECT 1
-#  pool:
-#    size: 5
 
 # swagger ui 환경설정
 springdoc:
